020421-1

+++ 7.Bien_Kieu_Du_Lieu_Va_Toan_Tu/Thuc_hanh/Su_Dung_Toan_Tu.js
@@ -2,7 +2,7 @@ let inputwidth;
 let inputheight;
 inputheight = prompt('Nhập chiều rộng');
 inputwidth = prompt('Nhập chiều dài');
-let width = parseInt(inputwidth);
-let height = parseInt(inputheight);
+let width = parseFloat(inputwidth);
+let height = parseFloat(inputheight);
 let area = width*height
 document.write("Diện tích hình chữ nhật là :" + area );
